host-a-model/qwen-spcs/agent/pyproject.toml
[project]
name = "snowflake-multi-agent-studio"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=1.1.0",
    "langchain-openai>=0.3.0",
    "langchain>=0.3.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.28.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "snowflake-snowpark-python>=1.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["setuptools>=68", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["agent*", "server*"]
